Make psuedocode V1, User class | Add function getRounds, createNewUser
I built out the initial psuedocode for the project.

I then made the class "User" to store information about the players
in the game.

I then got to work prototyping and testing some functions I believe
the finished product will need.

This includes a function to get the number of rounds of RPS the user
would like to play. Return should be a number.

This also includes a function to create a new player in the game.
Return should be a new instance of User class.